requirements.txt
PyYAML==6.0.1
Telethon==1.28.5
